Reorganize testing utilities
References #297
