+++ src/pybel/testing/__init__.py
@@ -1,0 +1,3 @@
+# -*- coding: utf-8 -*-
+
+"""Testing utilities for PyBEL."""
