+++ src/pybel/testing/resources/bel/Cytotoxic T-cell Signaling-2.0-Hs.json
