+++ src/pybel/testing/resources/bel/isolated.bel
