+++ src/pybel/testing/resources/bel/slushy.bel
