+++ src/pybel/testing/resources/bel/test_bel.bel
