+++ src/pybel/testing/resources/bel/thorough.bel
