+++ src/pybel/testing/resources/belanno/cell-line.belanno
