+++ src/pybel/testing/resources/belanno/confidence-1.0.0.belanno
