+++ src/pybel/testing/resources/belanno/mesh-diseases.belanno
