+++ src/pybel/testing/resources/belanno/species-taxonomy-id.belanno
