+++ src/pybel/testing/resources/belns/chebi.belns
