+++ src/pybel/testing/resources/belns/disease-ontology.belns
