+++ src/pybel/testing/resources/belns/go-biological-process.belns
