+++ src/pybel/testing/resources/belns/go-cellular-component.belns
