+++ src/pybel/testing/resources/belns/hgnc-human-genes.belns
