+++ src/pybel/testing/resources/belns/mesh-diseases.belns
