+++ src/pybel/testing/resources/belns/test_ns_1.belns
