+++ src/pybel/testing/resources/belns/test_ns_2.belns
