+++ src/pybel/testing/resources/belns/test_ns_empty.belns
